+++ scripts/requirements.txt
@@ -4,3 +4,4 @@ pyspark==2.4.3
 pandas==0.25.0
 scikit-learn==0.20.3
 scipy==1.3.0
+pytest==5.0.1
