+++ scripts/run.sh
@@ -1,4 +1,3 @@
 #!/usr/bin/env bash
 
-python3.7 -m pip install -r requirements.txt
-#spark-submit src/main.py
+PYTHONPATH=. python3.7 -m pip install -r requirements.txt
